feat: Move global TS types to their own file
